+++ src/plugin.rs
@@ -1,6 +1,4 @@
-use crate::assets::VelloAssetLoader;
 use crate::debug::DebugVisualizationsPlugin;
-use crate::player::LottiePlayerPlugin;
 use crate::render::VelloRenderPlugin;
 use crate::text::VelloFontLoader;
 use crate::{VelloAsset, VelloFont};
@@ -11,10 +9,15 @@ pub struct VelloPlugin;
 impl Plugin for VelloPlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins(VelloRenderPlugin)
-            .add_plugins((LottiePlayerPlugin, DebugVisualizationsPlugin))
+            .add_plugins(DebugVisualizationsPlugin)
             .init_asset::<VelloAsset>()
-            .init_asset_loader::<VelloAssetLoader>()
             .init_asset::<VelloFont>()
             .init_asset_loader::<VelloFontLoader>();
+        #[cfg(feature = "svg")]
+        app.add_plugins(crate::integrations::svg::SvgIntegrationPlugin);
+        #[cfg(feature = "lottie")]
+        app.add_plugins(crate::integrations::lottie::LottieIntegrationPlugin);
+        #[cfg(feature = "experimental-dotLottie")]
+        app.add_plugins(crate::integrations::dot_lottie::DotLottieIntegrationPlugin);
     }
 }